Moved command handling to a separate file
